pmt function fixed and working (maybe need to refactor later); solicitation show view created; loan_amount type changed to integer;

+++ db/migrate/20201028140609_create_solicitations.rb
@@ -1,7 +1,7 @@
 class CreateSolicitations < ActiveRecord::Migration[6.0]
   def change
     create_table :solicitations do |t|
-      t.float :loan_amount
+      t.integer :loan_amount
       t.integer :number_of_installments
       t.float :installment_amount
       t.boolean :pending, default: true
